% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sampling_interval.R
\name{avg_sampling_interval}
\alias{avg_sampling_interval}
\title{Find avg time between two sampling}
\usage{
avg_sampling_interval(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\value{
average time between in sampling in seconds
}
\description{
Find avg time between two sampling
}
\examples{

print("Hello!")

}
