% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collection_days.R
\name{data_collection_days}
\alias{data_collection_days}
\title{Find when data collection happened}
\usage{
data_collection_days(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\value{
date(s) when sampling occured
}
\description{
Find when data collection happened
}
\examples{

print("Hello!")

}
