% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_asd_visualisation.R
\name{raw_asd_viz}
\alias{raw_asd_viz}
\title{Visualise all raw spectra}
\usage{
raw_asd_viz(asd_data, include_wr = FALSE, alpha_value = NULL)
}
\arguments{
\item{asd_data}{ASD data as tibble}

\item{include_wr}{TRUE or FALSE to include White Ref in visualisation; default FALSE;}

\item{alpha_value}{by default set at 0.4 takes any value between 0 and 1}
}
\value{
graph of viz
}
\description{
Visualise all raw spectra
}
\examples{

print("Hello!")

}
