% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_asd_files.R
\name{read_asd_files}
\alias{read_asd_files}
\title{Read asd files from its parent directory}
\usage{
read_asd_files(file_path)
}
\arguments{
\item{file_path}{file path containing asd files}
}
\value{
a tibble or data frame
}
\description{
Read asd files from its parent directory
}
\examples{

print("Hello!")

}
