% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_type.R
\name{read_data_type}
\alias{read_data_type}
\title{Read the date when the asd file is created}
\usage{
read_data_type(fpath)
}
\arguments{
\item{fpath}{file path of asd file}
}
\value{
a character value of type of spectral data
}
\description{
Read the date when the asd file is created
}
\examples{

print("Hello!")

}
