% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_date_created.R
\name{read_date_created}
\alias{read_date_created}
\title{Read the date when the asd file is created}
\usage{
read_date_created(fpath)
}
\arguments{
\item{fpath}{file path of asd file}
}
\value{
a date object
}
\description{
Read the date when the asd file is created
}
\examples{

print("Hello!")

}
