% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_version.R
\name{read_file_version}
\alias{read_file_version}
\title{Reading file version from ASD file}
\usage{
read_file_version(fpath)
}
\arguments{
\item{fpath}{file path of asd file}
}
\value{
return a numeric value of file version number
}
\description{
Reading file version from ASD file
}
\examples{

print("Hello!")

}
