% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instrument.R
\name{read_instrument}
\alias{read_instrument}
\title{Reading instrument name from ASD file}
\usage{
read_instrument(fpath)
}
\arguments{
\item{fpath}{path to asd file}
}
\value{
an integer value of instrument serial number
}
\description{
Reading instrument name from ASD file
}
\examples{

print("Hello!")

}
