% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_integration_time.R
\name{read_integration_time}
\alias{read_integration_time}
\title{Read the integration time}
\usage{
read_integration_time(fpath)
}
\arguments{
\item{fpath}{file path of asd file}
}
\value{
a integer value denoting the type of spectral data
}
\description{
Read the integration time
}
\examples{

print("Hello!")

}
