% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spectra}
\alias{read_spectra}
\title{Read the spectra from ASD file}
\usage{
read_spectra(fpath)
}
\arguments{
\item{fpath}{file path of asd file}
}
\value{
a vector of reflectance
}
\description{
Read the spectra from ASD file
}
\examples{

print("Hello!")

}
