% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_splice1_wavelength.R
\name{read_splice1_wavelength}
\alias{read_splice1_wavelength}
\title{Get splice 1 wavelength}
\usage{
read_splice1_wavelength(fpath)
}
\arguments{
\item{fpath}{file path of asdfile}
}
\value{
wavelength of splice 1
}
\description{
Get splice 1 wavelength
}
\examples{

print("hello")

}
