% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spline2_wavelength.R
\name{read_splice2_wavelength}
\alias{read_splice2_wavelength}
\title{Splice 2 wavelength}
\usage{
read_splice2_wavelength(fpath)
}
\arguments{
\item{fpath}{file path containing asdfile}
}
\value{
wavelength of splice2
}
\description{
Splice 2 wavelength
}
