% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swir1_offset.R
\name{read_swir1_offset}
\alias{read_swir1_offset}
\title{SWIR1 offset}
\usage{
read_swir1_offset(fpath)
}
\arguments{
\item{fpath}{file path of asdfile}
}
\value{
swir1 offset
}
\description{
SWIR1 offset
}
