% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_swir2_offset.R
\name{read_swir2_offset}
\alias{read_swir2_offset}
\title{swir2 offset}
\usage{
read_swir2_offset(fpath)
}
\arguments{
\item{fpath}{file path of asdfile}
}
\value{
swir2 offset
}
\description{
swir2 offset
}
