% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_sampling_interval.R
\name{view_sampling_time}
\alias{view_sampling_time}
\title{Visualise ASD sampling interval}
\usage{
view_sampling_time(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\value{
a ggplot
}
\description{
Visualise ASD sampling interval
}
\examples{

print("Hello!")

}
