% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_white_ref.R
\name{view_white_ref}
\alias{view_white_ref}
\title{Visualise all or subset of white reference}
\usage{
view_white_ref(asd_data, which_white_ref = NULL)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}

\item{which_white_ref}{select ASD file number of the white ref (OPTIONAL)}
}
\value{
a ggplot
}
\description{
Visualise all or subset of white reference
}
\examples{

print("Hello!")

}
