% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_white_ref2.R
\name{view_white_ref2}
\alias{view_white_ref2}
\title{Visualise faceted White Reference}
\usage{
view_white_ref2(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\value{
a ggplot
}
\description{
Visualise faceted White Reference
}
\examples{

print("Hello!")

}
