% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_white_ref_quality.R
\name{view_white_ref_quality}
\alias{view_white_ref_quality}
\title{Visualise quality of white reference data}
\usage{
view_white_ref_quality(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\value{
a ggplot
}
\description{
Visualise quality of white reference data
}
\examples{

print("Hello!")

}
