% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr_sampling_interval.R
\name{wr_sampling_interval}
\alias{wr_sampling_interval}
\title{Average sampling between two white reference}
\usage{
wr_sampling_interval(asd_data)
}
\arguments{
\item{asd_data}{asd data read using read_asd_files() function}
}
\description{
Average sampling between two white reference
}
\examples{

print("Hello!")

}
